requirements/base.txt
Django==3.1.5
django-environ==0.4.5
django-redis==4.12.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.6
requests==2.25.1
